uploader/requirements.txt
requests
python-dotenv
python-magic
